Search the foundation component of the ICD-11

Search the foundation component of the ICD-11

Usage

icd_search_foundation(
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  q,
  subtree = NULL,
  chapter = NULL,
  flexisearch = FALSE,
  flat = TRUE,
  properties = NULL,
  release = NULL,
  highlight = FALSE,
  api_version = c("v2", "v1"),
  language = "en"
)

Arguments

base_url

The base URL of the API. Default uses the WHO API server at https://id.who.int. If you are using a locally deployed server or hosting your own ICD API server, you should specify the URL of your instance here.

client

The OAuth2 client produced through a call to icd_oauth_client().

scope

Scopes to be requested from the resource owner. Default is "icdapi_access" as specified in the ICD API documentation.

q

String. Text to be searched. Having the character % at the end will be regarded as a wild card for that word.

subtree

A string or vector of strings of URIs. If provided, the search will be performed on the entities provided and their descendants.

chapter

A string or vector of strings of chapter codes eg: c("01", "02") When provided, the search will be performed only on these chapters.

flexisearch

Logical. Default is FALSE. Changes the search mode to flexible search. In the regular search mode, the Coding Tool will only give you results that contain all of the words that you've used in your search. It accepts different variants or synonyms of the words but essentially it searches for a result that contains all components of your search. Whereas in flexible search mode, the results do not have to contain all of the words that are typed. It would still try to find the best matching phrase but there may be words in your search that are not matched at all. It is recommended to use flexible search only when regular search does not provide a result.

flat

Logical. Default is FALSE. If set to true the search result entities are provided in a nested data structure representing the ICD-11 hierarchy. Otherwise they are listed as flat list of matches.

properties

A string or a vector of strings for the properties to be searched. By default the system searches, "Title", "Synonyms", and "FullySpecifiedName". The valid values that could be used are: Title", "Synonym", "NarrowerTerm", "FullySpecifiedName", "Definition", and "Exclusion".

release

A string specifying the release version of the Foundation to search from. If not specified, defaults to the latest release version. See the available versions with icd_versions.

highlight

Logical. Default is FALSE. If set to FALSE the search result highlighting is turned off and the results don't contain special tags for highlighting where the results are found within the text.

api_version

Version of the API. Possible values are v1 or v2. For example, if you provide value v2, the API will respond in the format of the version 2 of the API. Default is v2.

language

ICD-API is multi-lingual. By changing this header, you may make the API respond in different languages. Languages will be available as the translations of ICD-11 completes. The values are language codes such as en, es, zh, etc. Depending on the release_id specified, the available languages will vary. Default is English ("en").

Value

An httr2_response object

Examples

icd_search_foundation(q = "cholera")
#> $error
#> [1] FALSE
#> 
#> $errorMessage
#> NULL
#> 
#> $resultChopped
#> [1] FALSE
#> 
#> $wordSuggestionsChopped
#> [1] FALSE
#> 
#> $guessType
#> [1] 2
#> 
#> $uniqueSearchId
#> [1] "8579e8dc-d00b-4cd7-823f-99c7f9b8bf19"
#> 
#> $words
#> NULL
#> 
#> $destinationEntities
#> $destinationEntities[[1]]
#> $destinationEntities[[1]]$id
#> [1] "http://id.who.int/icd/entity/257068234"
#> 
#> $destinationEntities[[1]]$title
#> [1] "Cholera"
#> 
#> $destinationEntities[[1]]$stemId
#> [1] "http://id.who.int/icd/entity/257068234"
#> 
#> $destinationEntities[[1]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[1]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs
#> $destinationEntities[[1]]$matchingPVs[[1]]
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$label
#> [1] "cholera syndrome"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$score
#> [1] 0.8780647
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/257068234"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$label
#> [1] "epidemic cholera"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$score
#> [1] 0.6680576
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/257068234"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]
#> $destinationEntities[[1]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$label
#> [1] "asiatic cholera"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$score
#> [1] 0.5068888
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/257068234"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[1]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[1]]$theCode
#> NULL
#> 
#> $destinationEntities[[1]]$score
#> [1] 1
#> 
#> $destinationEntities[[1]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[1]]$important
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[2]]
#> $destinationEntities[[2]]$id
#> [1] "http://id.who.int/icd/entity/733451521"
#> 
#> $destinationEntities[[2]]$title
#> [1] "Exposure to cholera"
#> 
#> $destinationEntities[[2]]$stemId
#> [1] "http://id.who.int/icd/entity/733451521"
#> 
#> $destinationEntities[[2]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[2]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$matchingPVs
#> $destinationEntities[[2]]$matchingPVs[[1]]
#> $destinationEntities[[2]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$label
#> [1] "cholera contact"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$score
#> [1] 0.8004876
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/733451521"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[2]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[2]]$theCode
#> NULL
#> 
#> $destinationEntities[[2]]$score
#> [1] 0.8004876
#> 
#> $destinationEntities[[2]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[3]]
#> $destinationEntities[[3]]$id
#> [1] "http://id.who.int/icd/entity/1471383174"
#> 
#> $destinationEntities[[3]]$title
#> [1] "Cholera vaccines"
#> 
#> $destinationEntities[[3]]$stemId
#> [1] "http://id.who.int/icd/entity/1471383174"
#> 
#> $destinationEntities[[3]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[3]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[3]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[3]]$theCode
#> NULL
#> 
#> $destinationEntities[[3]]$score
#> [1] 0.7708042
#> 
#> $destinationEntities[[3]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[4]]
#> $destinationEntities[[4]]$id
#> [1] "http://id.who.int/icd/entity/1205958647"
#> 
#> $destinationEntities[[4]]$title
#> [1] "Cholera due to Vibrio cholerae O1, biovar cholerae"
#> 
#> $destinationEntities[[4]]$stemId
#> [1] "http://id.who.int/icd/entity/1205958647"
#> 
#> $destinationEntities[[4]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[4]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$matchingPVs
#> $destinationEntities[[4]]$matchingPVs[[1]]
#> $destinationEntities[[4]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$label
#> [1] "classical cholera"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$score
#> [1] 0.7336125
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1205958647"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[4]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[4]]$theCode
#> NULL
#> 
#> $destinationEntities[[4]]$score
#> [1] 0.7336125
#> 
#> $destinationEntities[[4]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[5]]
#> $destinationEntities[[5]]$id
#> [1] "http://id.who.int/icd/entity/1474305701"
#> 
#> $destinationEntities[[5]]$title
#> [1] "Cholera carrier"
#> 
#> $destinationEntities[[5]]$stemId
#> [1] "http://id.who.int/icd/entity/1474305701"
#> 
#> $destinationEntities[[5]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[5]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[5]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[5]]$theCode
#> NULL
#> 
#> $destinationEntities[[5]]$score
#> [1] 0.6813468
#> 
#> $destinationEntities[[5]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[5]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[6]]
#> $destinationEntities[[6]]$id
#> [1] "http://id.who.int/icd/entity/824826401"
#> 
#> $destinationEntities[[6]]$title
#> [1] "Vibrio cholerae"
#> 
#> $destinationEntities[[6]]$stemId
#> [1] "http://id.who.int/icd/entity/824826401"
#> 
#> $destinationEntities[[6]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[6]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[6]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[6]]$theCode
#> NULL
#> 
#> $destinationEntities[[6]]$score
#> [1] 0.6482333
#> 
#> $destinationEntities[[6]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[6]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[7]]
#> $destinationEntities[[7]]$id
#> [1] "http://id.who.int/icd/entity/238837919"
#> 
#> $destinationEntities[[7]]$title
#> [1] "Antimonial cholera"
#> 
#> $destinationEntities[[7]]$stemId
#> [1] "http://id.who.int/icd/entity/238837919"
#> 
#> $destinationEntities[[7]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[7]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[7]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$chapter
#> [1] "22"
#> 
#> $destinationEntities[[7]]$theCode
#> NULL
#> 
#> $destinationEntities[[7]]$score
#> [1] 0.6280132
#> 
#> $destinationEntities[[7]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[7]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[8]]
#> $destinationEntities[[8]]$id
#> [1] "http://id.who.int/icd/entity/2053472774"
#> 
#> $destinationEntities[[8]]$title
#> [1] "Harmful effects of Cholera vaccine"
#> 
#> $destinationEntities[[8]]$stemId
#> [1] "http://id.who.int/icd/entity/2053472774"
#> 
#> $destinationEntities[[8]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[8]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[8]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$chapter
#> [1] "22"
#> 
#> $destinationEntities[[8]]$theCode
#> NULL
#> 
#> $destinationEntities[[8]]$score
#> [1] 0.4906102
#> 
#> $destinationEntities[[8]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[8]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[9]]
#> $destinationEntities[[9]]$id
#> [1] "http://id.who.int/icd/entity/581614179"
#> 
#> $destinationEntities[[9]]$title
#> [1] "Cholera due to Vibrio cholerae O1, biovar eltor"
#> 
#> $destinationEntities[[9]]$stemId
#> [1] "http://id.who.int/icd/entity/581614179"
#> 
#> $destinationEntities[[9]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[9]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs
#> $destinationEntities[[9]]$matchingPVs[[1]]
#> $destinationEntities[[9]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$label
#> [1] "cholera due to Cholera eltor"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$score
#> [1] 0.2167264
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/581614179"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]
#> $destinationEntities[[9]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$label
#> [1] "cholera due to Vibrio cholerae 01, Cholera eltor"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$score
#> [1] -0.5273059
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/581614179"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]
#> $destinationEntities[[9]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$label
#> [1] "cholera - vibrio cholerae 01 eltor biotype"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$score
#> [1] -0.9886038
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/581614179"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[9]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[9]]$theCode
#> NULL
#> 
#> $destinationEntities[[9]]$score
#> [1] 0.2167264
#> 
#> $destinationEntities[[9]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[9]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[9]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[10]]
#> $destinationEntities[[10]]$id
#> [1] "http://id.who.int/icd/entity/136299182"
#> 
#> $destinationEntities[[10]]$title
#> [1] "Cholera screening"
#> 
#> $destinationEntities[[10]]$stemId
#> [1] "http://id.who.int/icd/entity/136299182"
#> 
#> $destinationEntities[[10]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[10]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[10]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[10]]$theCode
#> NULL
#> 
#> $destinationEntities[[10]]$score
#> [1] 0.1900829
#> 
#> $destinationEntities[[10]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[10]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[11]]
#> $destinationEntities[[11]]$id
#> [1] "http://id.who.int/icd/entity/1472302478"
#> 
#> $destinationEntities[[11]]$title
#> [1] "Need for immunization against cholera alone"
#> 
#> $destinationEntities[[11]]$stemId
#> [1] "http://id.who.int/icd/entity/1472302478"
#> 
#> $destinationEntities[[11]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[11]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs
#> $destinationEntities[[11]]$matchingPVs[[1]]
#> $destinationEntities[[11]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$label
#> [1] "prophylactic vaccination against cholera"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$score
#> [1] 0.1540597
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1472302478"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[11]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[11]]$theCode
#> NULL
#> 
#> $destinationEntities[[11]]$score
#> [1] 0.1540597
#> 
#> $destinationEntities[[11]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[11]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[12]]
#> $destinationEntities[[12]]$id
#> [1] "http://id.who.int/icd/entity/1147064900"
#> 
#> $destinationEntities[[12]]$title
#> [1] "Cholera, live attenuated vaccines"
#> 
#> $destinationEntities[[12]]$stemId
#> [1] "http://id.who.int/icd/entity/1147064900"
#> 
#> $destinationEntities[[12]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[12]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[12]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[12]]$theCode
#> NULL
#> 
#> $destinationEntities[[12]]$score
#> [1] 0.1394327
#> 
#> $destinationEntities[[12]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[12]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[13]]
#> $destinationEntities[[13]]$id
#> [1] "http://id.who.int/icd/entity/1384028266"
#> 
#> $destinationEntities[[13]]$title
#> [1] "Enteritis due to Cholera due to Vibrio cholerae O1, biovar cholerae"
#> 
#> $destinationEntities[[13]]$stemId
#> [1] "http://id.who.int/icd/entity/1384028266"
#> 
#> $destinationEntities[[13]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[13]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[13]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$matchingPVs
#> $destinationEntities[[13]]$matchingPVs[[1]]
#> $destinationEntities[[13]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[13]]$matchingPVs[[1]]$label
#> [1] "Enteritis due to classical cholera"
#> 
#> $destinationEntities[[13]]$matchingPVs[[1]]$score
#> [1] -0.02692786
#> 
#> $destinationEntities[[13]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1384028266"
#> 
#> $destinationEntities[[13]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[13]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[13]]$theCode
#> NULL
#> 
#> $destinationEntities[[13]]$score
#> [1] -0.02692786
#> 
#> $destinationEntities[[13]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[13]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[13]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[13]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[14]]
#> $destinationEntities[[14]]$id
#> [1] "http://id.who.int/icd/entity/1859125126"
#> 
#> $destinationEntities[[14]]$title
#> [1] "Cholera, inactivated, whole cell vaccines"
#> 
#> $destinationEntities[[14]]$stemId
#> [1] "http://id.who.int/icd/entity/1859125126"
#> 
#> $destinationEntities[[14]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[14]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[14]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[14]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[14]]$theCode
#> NULL
#> 
#> $destinationEntities[[14]]$score
#> [1] -0.03419258
#> 
#> $destinationEntities[[14]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[14]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[14]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[14]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[14]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[15]]
#> $destinationEntities[[15]]$id
#> [1] "http://id.who.int/icd/entity/1972085775"
#> 
#> $destinationEntities[[15]]$title
#> [1] "Cholera due to Vibrio cholerae O139"
#> 
#> $destinationEntities[[15]]$stemId
#> [1] "http://id.who.int/icd/entity/1972085775"
#> 
#> $destinationEntities[[15]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[15]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[15]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[15]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[15]]$theCode
#> NULL
#> 
#> $destinationEntities[[15]]$score
#> [1] -0.0880754
#> 
#> $destinationEntities[[15]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[15]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[15]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[15]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[15]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[16]]
#> $destinationEntities[[16]]$id
#> [1] "http://id.who.int/icd/entity/21927667"
#> 
#> $destinationEntities[[16]]$title
#> [1] "Vibrio cholerae O139"
#> 
#> $destinationEntities[[16]]$stemId
#> [1] "http://id.who.int/icd/entity/21927667"
#> 
#> $destinationEntities[[16]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[16]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[16]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[16]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[16]]$theCode
#> NULL
#> 
#> $destinationEntities[[16]]$score
#> [1] -0.0880754
#> 
#> $destinationEntities[[16]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[16]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[16]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[16]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[16]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[17]]
#> $destinationEntities[[17]]$id
#> [1] "http://id.who.int/icd/entity/1388151503"
#> 
#> $destinationEntities[[17]]$title
#> [1] "Unintentional exposure to or harmful effects of cholera vaccine"
#> 
#> $destinationEntities[[17]]$stemId
#> [1] "http://id.who.int/icd/entity/1388151503"
#> 
#> $destinationEntities[[17]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[17]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[17]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[17]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$chapter
#> [1] "23"
#> 
#> $destinationEntities[[17]]$theCode
#> NULL
#> 
#> $destinationEntities[[17]]$score
#> [1] -0.2044832
#> 
#> $destinationEntities[[17]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[17]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[17]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[17]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[17]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[18]]
#> $destinationEntities[[18]]$id
#> [1] "http://id.who.int/icd/entity/194483911"
#> 
#> $destinationEntities[[18]]$title
#> [1] "Vibrio cholerae O1, biovar cholerae"
#> 
#> $destinationEntities[[18]]$stemId
#> [1] "http://id.who.int/icd/entity/194483911"
#> 
#> $destinationEntities[[18]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[18]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[18]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[18]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[18]]$theCode
#> NULL
#> 
#> $destinationEntities[[18]]$score
#> [1] -0.2060272
#> 
#> $destinationEntities[[18]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[18]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[18]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[18]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[18]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[19]]
#> $destinationEntities[[19]]$id
#> [1] "http://id.who.int/icd/entity/1832869108"
#> 
#> $destinationEntities[[19]]$title
#> [1] "Need for immunization against cholera with typhoid-paratyphoid"
#> 
#> $destinationEntities[[19]]$stemId
#> [1] "http://id.who.int/icd/entity/1832869108"
#> 
#> $destinationEntities[[19]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[19]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[19]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$matchingPVs
#> $destinationEntities[[19]]$matchingPVs[[1]]
#> $destinationEntities[[19]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[19]]$matchingPVs[[1]]$label
#> [1] "cholera with TAB immunisation"
#> 
#> $destinationEntities[[19]]$matchingPVs[[1]]$score
#> [1] -0.3055559
#> 
#> $destinationEntities[[19]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1832869108"
#> 
#> $destinationEntities[[19]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]
#> $destinationEntities[[19]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]$label
#> [1] "prophylactic vaccination against cholera with typhoid-paratyphoid"
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]$score
#> [1] -1.877626
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1832869108"
#> 
#> $destinationEntities[[19]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[19]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[19]]$theCode
#> NULL
#> 
#> $destinationEntities[[19]]$score
#> [1] -0.3055559
#> 
#> $destinationEntities[[19]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[19]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[19]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[19]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[20]]
#> $destinationEntities[[20]]$id
#> [1] "http://id.who.int/icd/entity/304914281"
#> 
#> $destinationEntities[[20]]$title
#> [1] "Intentional self-harm by exposure to or harmful effects of cholera vaccine"
#> 
#> $destinationEntities[[20]]$stemId
#> [1] "http://id.who.int/icd/entity/304914281"
#> 
#> $destinationEntities[[20]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[20]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[20]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[20]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$chapter
#> [1] "23"
#> 
#> $destinationEntities[[20]]$theCode
#> NULL
#> 
#> $destinationEntities[[20]]$score
#> [1] -0.3677743
#> 
#> $destinationEntities[[20]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[20]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[20]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[20]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[20]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[21]]
#> $destinationEntities[[21]]$id
#> [1] "http://id.who.int/icd/entity/1642936749"
#> 
#> $destinationEntities[[21]]$title
#> [1] "Exposure to or harmful effects of undetermined intent of cholera vaccine"
#> 
#> $destinationEntities[[21]]$stemId
#> [1] "http://id.who.int/icd/entity/1642936749"
#> 
#> $destinationEntities[[21]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[21]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[21]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[21]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$chapter
#> [1] "23"
#> 
#> $destinationEntities[[21]]$theCode
#> NULL
#> 
#> $destinationEntities[[21]]$score
#> [1] -0.3853351
#> 
#> $destinationEntities[[21]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[21]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[21]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[21]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[21]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[22]]
#> $destinationEntities[[22]]$id
#> [1] "http://id.who.int/icd/entity/402305888"
#> 
#> $destinationEntities[[22]]$title
#> [1] "Assault by or harmful exposure to cholera vaccine"
#> 
#> $destinationEntities[[22]]$stemId
#> [1] "http://id.who.int/icd/entity/402305888"
#> 
#> $destinationEntities[[22]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[22]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[22]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[22]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$chapter
#> [1] "23"
#> 
#> $destinationEntities[[22]]$theCode
#> NULL
#> 
#> $destinationEntities[[22]]$score
#> [1] -0.4972295
#> 
#> $destinationEntities[[22]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[22]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[22]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[22]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[22]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[23]]
#> $destinationEntities[[23]]$id
#> [1] "http://id.who.int/icd/entity/375406584"
#> 
#> $destinationEntities[[23]]$title
#> [1] "Enteritis due to cholera due to Vibrio cholerae O1, biovar eltor"
#> 
#> $destinationEntities[[23]]$stemId
#> [1] "http://id.who.int/icd/entity/375406584"
#> 
#> $destinationEntities[[23]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[23]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[23]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$matchingPVs
#> $destinationEntities[[23]]$matchingPVs[[1]]
#> $destinationEntities[[23]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[23]]$matchingPVs[[1]]$label
#> [1] "Enteritis due to cholera due to Cholera eltor"
#> 
#> $destinationEntities[[23]]$matchingPVs[[1]]$score
#> [1] -0.5438139
#> 
#> $destinationEntities[[23]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/375406584"
#> 
#> $destinationEntities[[23]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]
#> $destinationEntities[[23]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]$label
#> [1] "infectious enteritis due to vibrio cholerae 01, biovar eltor"
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]$score
#> [1] -2.08133
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/375406584"
#> 
#> $destinationEntities[[23]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[23]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[23]]$theCode
#> NULL
#> 
#> $destinationEntities[[23]]$score
#> [1] -0.5438139
#> 
#> $destinationEntities[[23]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[23]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[23]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[23]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[24]]
#> $destinationEntities[[24]]$id
#> [1] "http://id.who.int/icd/entity/802386629"
#> 
#> $destinationEntities[[24]]$title
#> [1] "Vibrio cholerae O1, biovar eltor"
#> 
#> $destinationEntities[[24]]$stemId
#> [1] "http://id.who.int/icd/entity/802386629"
#> 
#> $destinationEntities[[24]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[24]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[24]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[24]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[24]]$theCode
#> NULL
#> 
#> $destinationEntities[[24]]$score
#> [1] -0.9893008
#> 
#> $destinationEntities[[24]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[24]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[24]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[24]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[24]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[25]]
#> $destinationEntities[[25]]$id
#> [1] "http://id.who.int/icd/entity/1170831944"
#> 
#> $destinationEntities[[25]]$title
#> [1] "Enteritis due to cholera due to Vibrio cholerae, non-O1 strains"
#> 
#> $destinationEntities[[25]]$stemId
#> [1] "http://id.who.int/icd/entity/1170831944"
#> 
#> $destinationEntities[[25]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[25]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[25]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[25]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[25]]$theCode
#> NULL
#> 
#> $destinationEntities[[25]]$score
#> [1] -1.013056
#> 
#> $destinationEntities[[25]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[25]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[25]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[25]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[25]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[26]]
#> $destinationEntities[[26]]$id
#> [1] "http://id.who.int/icd/entity/1322480936"
#> 
#> $destinationEntities[[26]]$title
#> [1] "Cholera, combinations with typhoid vaccine, inactivated, whole cell vaccines"
#> 
#> $destinationEntities[[26]]$stemId
#> [1] "http://id.who.int/icd/entity/1322480936"
#> 
#> $destinationEntities[[26]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[26]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[26]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[26]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$chapter
#> [1] "X"
#> 
#> $destinationEntities[[26]]$theCode
#> NULL
#> 
#> $destinationEntities[[26]]$score
#> [1] -1.258335
#> 
#> $destinationEntities[[26]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[26]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[26]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[26]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[26]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[27]]
#> $destinationEntities[[27]]$id
#> [1] "http://id.who.int/icd/entity/296398626"
#> 
#> $destinationEntities[[27]]$title
#> [1] "Drugs medicaments or biological substances associated with injury or harm in therapeutic use, Bacterial vaccines, Cholera vaccine"
#> 
#> $destinationEntities[[27]]$stemId
#> [1] "http://id.who.int/icd/entity/296398626"
#> 
#> $destinationEntities[[27]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[27]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[27]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[27]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$chapter
#> [1] "23"
#> 
#> $destinationEntities[[27]]$theCode
#> NULL
#> 
#> $destinationEntities[[27]]$score
#> [1] -1.326383
#> 
#> $destinationEntities[[27]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[27]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[27]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[27]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[27]]$descendants
#> list()
#> 
#> 
#>