Get information on various ICD 11 Foundation entities

Get information on various ICD 11 Foundation entities

Usage

icd_get_foundation(
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  parse = TRUE
)

icd_get_entity(
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  id,
  release = NULL,
  include = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  parse = TRUE
)

Arguments

base_url

The base URL of the API. Default uses the WHO API server at https://id.who.int. If you are using a locally deployed server or hosting your own ICD API server, you should specify the URL of your instance here.

client

The OAuth2 client produced through a call to icd_oauth_client().

scope

Scopes to be requested from the resource owner. Default is "icdapi_access" as specified in the ICD API documentation.

release

A string specifying the release version of the Foundation to search from. If not specified, defaults to the latest release version. See the available versions with icd_versions.

api_version

Version of the API. Possible values are v1 or v2. For example, if you provide value v2, the API will respond in the format of the version 2 of the API. Default is v2.

language

ICD-API is multi-lingual. By changing this header, you may make the API respond in different languages. Languages will be available as the translations of ICD-11 completes. The values are language codes such as en, es, zh, etc. Depending on the release_id specified, the available languages will vary. Default is English ("en").

parse

Logical. Should JSON response body be parsed? Default is TRUE. If FALSE, response body is kept as raw JSON.

id

Unique numerical identifier for an entity.

include

A string or a vector of strings for optional property values to be included in the response. The property values that can be included are "ancestor", "descendant", or "diagnosticCriteria". If not specified, these properties are not included in the response.

Value

A list or JSON (depending on parse) with information on specified ICD 11 Foundation and top level entities

Examples

icd_get_foundation()
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForTopLevel.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/entity"
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "WHO Family of International Classifications Foundation"
#> 
#> 
#> $releaseId
#> [1] "2024-01"
#> 
#> $availableLanguages
#> $availableLanguages[[1]]
#> [1] "ar"
#> 
#> $availableLanguages[[2]]
#> [1] "cs"
#> 
#> $availableLanguages[[3]]
#> [1] "en"
#> 
#> $availableLanguages[[4]]
#> [1] "es"
#> 
#> $availableLanguages[[5]]
#> [1] "fr"
#> 
#> $availableLanguages[[6]]
#> [1] "pt"
#> 
#> $availableLanguages[[7]]
#> [1] "ru"
#> 
#> $availableLanguages[[8]]
#> [1] "tr"
#> 
#> $availableLanguages[[9]]
#> [1] "uz"
#> 
#> $availableLanguages[[10]]
#> [1] "zh"
#> 
#> 
#> $releaseDate
#> [1] "2024-01-21"
#> 
#> $child
#> $child[[1]]
#> [1] "http://id.who.int/icd/entity/448895267"
#> 
#> $child[[2]]
#> [1] "http://id.who.int/icd/entity/1405434703"
#> 
#> 
#> $browserUrl
#> [1] "https://icd.who.int/browse/2024-01/foundation/en"
#> 
icd_get_entity(id = "1112016389")
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForFoundationEntity.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/entity/1112016389"
#> 
#> $parent
#> $parent[[1]]
#> [1] "http://id.who.int/icd/entity/762120574"
#> 
#> 
#> $child
#> $child[[1]]
#> [1] "http://id.who.int/icd/entity/1181462425"
#> 
#> 
#> $browserUrl
#> [1] "https://icd.who.int/browse/2024-01/foundation/en#1112016389"
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "Gonorrhoea of penis"
#> 
#> 
#> $relatedEntitiesInMaternalChapter
#> $relatedEntitiesInMaternalChapter[[1]]
#> [1] "http://id.who.int/icd/entity/1477699562"
#> 
#> 
#> $relatedEntitiesInPerinatalChapter
#> $relatedEntitiesInPerinatalChapter[[1]]
#> [1] "http://id.who.int/icd/entity/911707612"
#> 
#>