Search the foundation component of the ICD-11

Search the foundation component of the ICD-11

Usage

icd_search_foundation(
  q,
  subtree = NULL,
  chapter = NULL,
  flexisearch = FALSE,
  flat = TRUE,
  properties = NULL,
  release = NULL,
  highlight = FALSE,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access"
)

Arguments

q

String. Text to be searched. Having the character % at the end will be regarded as a wild card for that word.

subtree

A string or vector of strings of URIs. If provided, the search will be performed on the entities provided and their descendants.

chapter

A string or vector of strings of chapter codes eg: c("01", "02") When provided, the search will be performed only on these chapters.

flexisearch

Logical. Default is FALSE. Changes the search mode to flexible search. In the regular search mode, the Coding Tool will only give you results that contain all of the words that you've used in your search. It accepts different variants or synonyms of the words but essentially it searches for a result that contains all components of your search. Whereas in flexible search mode, the results do not have to contain all of the words that are typed. It would still try to find the best matching phrase but there may be words in your search that are not matched at all. It is recommended to use flexible search only when regular search does not provide a result.

flat

Logical. Default is FALSE. If set to true the search result entities are provided in a nested data structure representing the ICD-11 hierarchy. Otherwise they are listed as flat list of matches.

properties

A string or a vector of strings for the properties to be searched. By default the system searches, "Title", "Synonyms", and "FullySpecifiedName". The valid values that could be used are: "Title", "Synonym", "NarrowerTerm", "FullySpecifiedName", "Definition", and "Exclusion".

release

A string specifying the release version of the Foundation to search from. If not specified, defaults to the latest release version. See the available versions with icd_versions.

highlight

Logical. Default is FALSE. If set to FALSE the search result highlighting is turned off and the results don't contain special tags for highlighting where the results are found within the text.

api_version

Version of the API. Possible values are v1 or v2. For example, if you provide value v2, the API will respond in the format of the version 2 of the API. Default is v2.

language

ICD-API is multi-lingual. By changing this header, you may make the API respond in different languages. Languages will be available as the translations of ICD-11 completes. The values are language codes such as en, es, zh, etc. Depending on the release_id specified, the available languages will vary. Default is English ("en").

tabular

Logical. Should output be structured into a tibble? Default to TRUE.

verbose

Logical. Should non-warning and non-error messages be printed? Default is TRUE.

base_url

The base URL of the API. Default uses the WHO API server at https://id.who.int. If you are using a locally deployed server or hosting your own ICD API server, you should specify the URL of your instance here.

client

The OAuth2 client produced through a call to icd_oauth_client().

scope

Scopes to be requested from the resource owner. Default is "icdapi_access" as specified in the ICD API documentation.

Value

A tibble of search results.

Examples

icd_search_foundation("colorectal cancer")
#> Release `2024-01` matches a known release for ICD-11.
#> Language `en` is available for the release specified.
#> # A tibble: 8 × 18
#>   id                    title stemId isLeaf postcoordinationAvai…¹ hasCodingNote
#> * <chr>                 <chr> <chr>  <lgl>                   <int> <lgl>        
#> 1 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 2 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 3 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 4 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 5 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 6 http://id.who.int/ic… Mali… http:… FALSE                       0 FALSE        
#> 7 http://id.who.int/ic… Fami… http:… FALSE                       0 FALSE        
#> 8 http://id.who.int/ic… Here… http:… FALSE                       0 FALSE        
#> # ℹ abbreviated name: ¹​postcoordinationAvailability
#> # ℹ 12 more variables: hasMaternalChapterLink <lgl>,
#> #   hasPerinatalChapterLink <lgl>, matchingPVs <named list>,
#> #   propertiesTruncated <lgl>, isResidualOther <lgl>,
#> #   isResidualUnspecified <lgl>, chapter <chr>, score <dbl>,
#> #   titleIsASearchResult <lgl>, titleIsTopScore <lgl>, entityType <int>,
#> #   important <lgl>